Fixed libdrat2er.a output name.
The CMake target name for the drat2er static library is "lib_drat2er",
since "drat2er" is the CMake target name for the drat2er binary.

Thus, the static library's archive file was being called
"liblib_drat2er.a" on linuz, which is a bit unfortunate. Now, after this
property change, it is called "libdrat2er.a", which is what one would
expect.